+++ CHANGES.md
@@ -10,6 +10,11 @@ Previously, only objects and input objects could be documented.
 New function `com.walmartlabs.lacinia.util/inject-resolvers` is an alternate way
 to attach resolvers to a schema.
 
+It is now possible to combine external documentation, from a Markdown file,
+into an EDN schema.
+See `com.walmartlabs.lacinia.parser.docs/parse-docs` and
+`com.walmartlabs.lacinia.util/inject-descriptions`.
+
 ## 0.26.0 -- 20 Apr 2018
 
 Lacinia now supports the `:roots` key in the input schema, which makes